Completed rewrite
Functionally now is equivalent to the original
version

WIP #1107
