+++ jwst/lib/set_telescope_pointing.py
@@ -1,230 +1,221 @@
-#!/usr/bin/env python
-
-# Copyright (C) 2010-2011 Association of Universities for Research in Astronomy (AURA)
-
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-
-#     1. Redistributions of source code must retain the above copyright
-#       notice, this list of conditions and the following disclaimer.
-
-#     2. Redistributions in binary form must reproduce the above
-#       copyright notice, this list of conditions and the following
-#       disclaimer in the documentation and/or other materials provided
-#       with the distribution.
-
-#     3. The name of AURA and its representatives may not be used to
-#       endorse or promote products derived from this software without
-#       specific prior written permission.
-
-# THIS SOFTWARE IS PROVIDED BY AURA ``AS IS'' AND ANY EXPRESS OR IMPLIED
-# WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
-# MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL AURA BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-# OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
-# TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
-# USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-# DAMAGE.
-
-'''
-This script adds absolute pointing information to the FITS files provided
-to it on the command line (one or more).
-
-Currently it only uses a constant value for the engineering keywords
-since the Engineering Database does not yet contain them.
-
-It assumes the following keywords are present in the file header:
-
-V2_REF (arcseconds)
-V3_REF (arcseconds)
-VPARITY (+1 or -1)
-V3I_YANG (decimal degrees)
-
-The keywords added are:
-
-RA_V1
-DEC_V1
-PA_V3
-CRVAL1
-CRVAL2
-PC1_1
-PC1_2
-PC2_1
-PC2_2
-
-It does not currently place the new keywords in any particular location
-in the header other than what is required by the standard.
-'''
-
-from __future__ import print_function, division
-
-from collections import namedtuple
+"""Set Telescope Pointing from quaternions"""
 import logging
-import sys
-
-import astropy.io.fits as fits
 import numpy as np
-from numpy import cos, sin
+from numpy import (cos, sin)
+
+from astropy.io import fits
+from namedlist import namedlist
+
 from jwst.lib.engdb_tools import (
     ENGDB_BASE_URL,
     ENGDB_Service,
 )
 
+# Setup logging
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
-handler = logging.StreamHandler()
-handler.setLevel(logging.DEBUG)
-logger.addHandler(handler)
+logger.addHandler(logging.NullHandler())
 
-# Define the return from get_pointing
-Pointing_Quaternions = namedtuple(
-    'Pointing_Quaternions',
-    ['q', 'j2fgs_matrix', 'fsmcorr', 'obstime']
+# ################################
+# Default transformation matricies
+# ################################
+FGS12SIFOV_DEFAULT = np.array(
+    [[0.9999994955442, 0.0000000000000, 0.0010044457459],
+     [0.0000011174826, 0.9999993811310, -0.0011125359826],
+     [-0.0010044451243, 0.0011125365439, 0.9999988766756]]
 )
 
+J2FGS_MATRIX_DEFAULT = np.array(
+    [[0.999997425983907, 0, -0.002268926080840],
+     [0., 1., 0.],
+     [0.002268926080840, 0., 0.999997425983907]]
+)
 
-def add_wcs(filename):
-    '''
-    Given the name of a valid partially populated level 1b JWST file,
-    determine the simple WCS parameters from the SIAF keywords in that
-    file and the engineering parameters that contain information about
-    the telescope pointing.
+SIFOV2V_DEFAULT = np.array(
+    [[0.99999742598, 0., 0.00226892608],
+     [0., 1., 0.],
+     [-0.00226892608, 0., 0.99999742598]]
+)
 
-    It presumes all the accessed keywords are present (see first block).
-    '''
-    hdulist = fits.open(filename, 'update')
-    pheader = hdulist[0].header
-    obsstart = float(pheader['EXPSTART'])
-    obsend = float(pheader['EXPEND'])
-    v2ref = float(pheader['V2_REF'])
-    v3ref = float(pheader['V3_REF'])
-    v3idlyang = float(pheader['V3I_YANG'])
-    vparity = int(pheader['VPARITY'])
-
-    # ##########################################
-    # WARNINGWARNINGWARNINGWARNINGWARNINGWARNING
-    #
-    # Get engineering parameters about scope pointing.
-    # In normal operations, if the paramters cannot be found
-    # this should fail.
-    # However, for prelaunch, we'll dummy out.
-    try:
-        q, j2fgs_matrix, fsmcorr, obstime = get_pointing(obsstart, obsend)
-    except ValueError as exception:
-        ra = pheader['TARG_RA']
-        dec = pheader['TARG_DEC']
-        roll = 0
-
-        logger.warning(
-            'Cannot retrieve telescope pointing.'
-            '\n{}'
-            '\nUsing TARG_RA={}, TARG_DEC={} and PA_V3=0 '
-            'to set pointing.'.format(exception, ra, dec)
-        )
-
-        local_roll = compute_local_roll(roll, ra, dec, v2ref, v3ref)
-        wcsinfo = (ra, dec, local_roll)
-        vinfo = (ra, dec, roll)
-        crval1, crval2, pa_aper_deg = wcsinfo
-        v1_ra_deg, v1_dec_deg, v3_pa_deg = vinfo
-        pa_aper_deg = local_roll - vparity * v3idlyang
-    else:
-        # compute relevant WCS information
-        logger.info('Successful read of engineering quaternions.')
-        logger.debug('q={}'.format(q))
-        logger.debug('j2fgs_matrix={}'.format(j2fgs_matrix))
-        logger.debug('fsmcorr={}'.format(fsmcorr))
-        wcsinfo, vinfo = calc_wcs(v2ref, v3ref, v3idlyang, vparity,
-                                  q, j2fgs_matrix, fsmcorr)
-        crval1, crval2, pa_aper_deg = wcsinfo
-        v1_ra_deg, v1_dec_deg, v3_pa_deg = vinfo
-        local_roll = compute_local_roll(v3_pa_deg, crval1, crval2, v2ref, v3ref)
-
-        logger.info(
-            'Computed coordinates from quaternions:'
-            '\n\tRA = {} DEC={} PA_V3={}'.format(crval1, crval2, v3_pa_deg)
-        )
-
-    pheader['RA_V1'] = v1_ra_deg
-    pheader['DEC_V1'] = v1_dec_deg
-    pheader['PA_V3'] = v3_pa_deg
-    pheader['CRVAL1'] = crval1
-    pheader['CRVAL2'] = crval2
-    pheader['PC1_1'] = -np.cos(pa_aper_deg * D2R)
-    pheader['PC1_2'] = np.sin(pa_aper_deg * D2R)
-    pheader['PC2_1'] = np.sin(pa_aper_deg * D2R)
-    pheader['PC2_2'] = np.cos(pa_aper_deg * D2R)
-    pheader['RA_REF'] = crval1
-    pheader['DEC_REF'] = crval2
-    pheader['ROLL_REF'] = local_roll
-    pheader['WCSAXES'] = len(pheader['CTYPE*'])
-    hdulist.flush()
-    hdulist.close()
-    logger.info('WCS info for {} complete.'.format(filename))
-
-
-def m_v_to_siaf(ya, v3, v2, vidlparity):  # This is a 321 rotation
-    mat = np.array([[cos(v3)*cos(v2),
-                    cos(v3)*sin(v2),
-                    sin(v3)],
-                   [-cos(ya)*sin(v2)+sin(ya)*sin(v3)*cos(v2),
-                    cos(ya)*cos(v2)+sin(ya)*sin(v3)*sin(v2),
-                    -sin(ya)*cos(v3)],
-                   [-sin(ya)*sin(v2)-cos(ya)*sin(v3)*cos(v2),
-                    sin(ya)*cos(v2)-cos(ya)*sin(v3)*sin(v2),
-                    cos(ya)*cos(v3)]])
-    pmat = np.array([[0., vidlparity, 0.],
-                     [0., 0., 1.],
-                     [1., 0., 0.]])
-    return np.dot(pmat, mat)
-
-
-def vector_to_ra_dec(v):
-    """Returns tuple of spherical angles from unit direction Vector """
-    ra = np.arctan2(v[1], v[0])
-    dec = np.arcsin(v[2])
-    if ra < 0.:
-        ra += 2. * np.pi
-    return(ra, dec)
-
+# Degree, radian, angle transformations
 R2D = 180./np.pi
 D2R = np.pi/180.
 A2R = D2R/3600.
 R2A = 3600.*R2D
 
-# Define the FGS1 to SI-FOV DCM,  Transpose of DCM in SE-20 section 5.8.4.2.
+# SIAF container
+SIAF = namedlist(
+    'SIAF',
+    ['v2ref', 'v3ref', 'v3idlyang', 'vparity'],
+    default=None
+)
 
-m1 = np.array(
-    [[0.9999994955442, 0.0000000000000, 0.0010044457459],
-     [0.0000011174826, 0.9999993811310, -0.0011125359826],
-     [-0.0010044451243, 0.0011125365439, 0.9999988766756]])
-m2 = np.array(
-    [[0, 0, 1],
-     [1, 0, 0],
-     [0, 1, 0]])
+# Pointing container
+Pointing = namedlist(
+    'Pointing',
+    ['q', 'j2fgs_matrix', 'fsmcorr', 'obstime'],
+    default=None
+)
 
-m_fgs1_to_sifov = np.dot(m2, m1)
-m_fgs1_to_sifovT = m_fgs1_to_sifov.transpose()
+# Transforms
+Transforms = namedlist(
+    'Transforms',
+    [
+        'm_eci2j',            # ECI to J-Frame
+        'm_j2fgs1',           # J-Frame to FGS1
+        'm_sifov_fsm_delta',  # FSM correction
+        'm_fgs12sifov',       # FGS1 to SIFOV
+        'm_eci2sifov',        # ECI to SIFOV
+        'm_sifov2v',          # SIFOV to V1
+        'm_eci2v',            # ECI to V
+        'm_v2siaf',           # V to SIAF
+        'm_eci2siaf'          # ECI to SIAF
+    ],
+    default=None
+)
 
-# Define the SI-FOV to V-frame DCM,  From Luis' IOC
-m_sifov_to_v = np.array(
-    [[0.99999742598, 0., 0.00226892608],
-     [0., 1., 0.],
-     [-0.00226892608, 0., 0.99999742598]])
+# WCS reference container
+WCSRef = namedlist(
+    'WCSRef',
+    ['ra', 'dec', 'pa'],
+    default=None
+)
 
 
-def calc_wcs(v2ref, v3ref, v3idlyang, vidlparity,
-             q, j2fgs_matrix, fsmcorr):
-    '''
-    v2ref (arcsec), v3ref (arcsec), v3idlyang (deg), vidlparity (+1 or -1),
-    are the relevant siaf parameters. The assumed units are shown in
-    parentheses.
+def add_wcs(filename, default_pa_v3=0.):
+    """Add WCS information to a FITS file
+
+    Telescope orientation is attempted to be obtained from
+    the engineering database. Failing that, a default pointing
+    is used based on proposal target.
+
+    The FITS file is updated in-place.
+
+    Parameters
+    ----------
+    filename: str
+        The path to a FITS file
+
+    default_pa_v3: float
+        The V3 position angle to use if the pointing information
+        is not found.
+    """
+    hdul = fits.open(filename, mode='update')
+    primary_header = hdul[0].header
+    update_header_wcs(primary_header, default_pa_v3=default_pa_v3)
+    hdul.flush()
+    hdul.close()
+
+
+def update_header_wcs(header, default_pa_v3=0.):
+    """Update WCS pointing information
+
+    Given a FITS Header Data Unit, determine the simple WCS parameters
+    from the SIAF keywords in the HDU and the engineering parameters
+    that contain information about the telescope pointing.
+
+    It presumes all the accessed keywords are present (see first block).
+
+    Parameters
+    ----------
+    header: dict-like
+        The header to update. Most likely
+        a FITS header such as
+        `astropy.io.fits.ImageHDU.header`
+
+    default_pa_v3: float
+        If pointing information cannot be retrieved,
+        use this as the V3 position angle.
+    """
+
+    # Get the SIAF and observation parameters
+    obsstart = float(header['EXPSTART'])
+    obsend = float(header['EXPEND'])
+    siaf = SIAF(
+        v2ref=float(header['V2_REF']),
+        v3ref=float(header['V3_REF']),
+        v3idlyang=float(header['V3I_YANG']),
+        vparity=int(header['VPARITY'])
+    )
+
+    # Get the pointing information
+    try:
+        pointing = get_pointing(obsstart, obsend)
+    except ValueError as exception:
+        ra = header['TARG_RA']
+        dec = header['TARG_DEC']
+        roll = default_pa_v3
+
+        logger.warning(
+            'Cannot retrieve telescope pointing.'
+            '\n{}'
+            '\nUsing TARG_RA={}, TARG_DEC={} and PA_V3={} '
+            'to set pointing.'.format(exception, ra, dec, roll)
+        )
+
+        wcsinfo = WCSRef(ra, dec, roll)
+        vinfo = WCSRef(ra, dec, roll)
+
+    else:
+        # compute relevant WCS information
+        logger.info('Successful read of engineering quaternions:')
+        logger.info('\tPointing = {}'.format(pointing))
+        wcsinfo, vinfo = calc_wcs(pointing, siaf)
+
+    logger.info('Aperture WCS info: {}'.format(wcsinfo))
+    logger.info('V1 WCS info: {}'.format(vinfo))
+
+    # Update V1 pointing
+    header['RA_V1'] = vinfo.ra
+    header['DEC_V1'] = vinfo.dec
+    header['PA_V3'] = vinfo.pa
+
+    # Update Aperture pointing
+    header['CRVAL1'] = wcsinfo.ra
+    header['CRVAL2'] = wcsinfo.dec
+    header['PC1_1'] = -np.cos(wcsinfo.pa * D2R)
+    header['PC1_2'] = np.sin(wcsinfo.pa * D2R)
+    header['PC2_1'] = np.sin(wcsinfo.pa * D2R)
+    header['PC2_2'] = np.cos(wcsinfo.pa * D2R)
+    header['RA_REF'] = wcsinfo.ra
+    header['DEC_REF'] = wcsinfo.dec
+    header['ROLL_REF'] = compute_local_roll(
+        vinfo.pa, wcsinfo.ra, wcsinfo.dec, siaf.v2ref, siaf.v3ref
+    )
+    header['WCSAXES'] = len(header['CTYPE*'])
+
+    logger.info('Header successfully updated')
+
+
+def calc_wcs(pointing, siaf):
+    """Transform from the given SIAF information and Pointing
+    the aperture and V1 wcs
+
+    Parameters
+    ----------
+    siaf: SIAF
+        The SIAF transformation. See ref:`Notes` for further details
+    pointing: Pointing
+        The telescope pointing. See ref:`Notes` for further details
+
+    Returns
+    -------
+    (wcsinfo, vinfo): (WCSRef, WCSRef)
+        A 2-tuple is returned with the WCS pointing for
+        the aperture and the V1 axis
+
+    Notes
+    -----
+
+    The SIAF information is as follows:
+
+    v2ref (arcsec), v3ref (arcsec), v3idlyang (deg), vidlparity
+    (+1 or -1), are the relevant siaf parameters. The assumed
+    units are shown in parentheses.
 
     It is assumed that the siaf ref position is the corresponding WCS
     reference position.
+
+    The `Pointing` information is as follows:
 
     Parameter q is the SA_ZATTEST<n> engineering parameters where
     n ranges from 1 to 4.
@@ -236,87 +227,378 @@ def calc_wcs(v2ref, v3ref, v3idlyang, vidlparity,
 
     Parameter fsmcorr are two values provided as a list consisting of:
     [SA_ZADUCMDX, SA_ZADUCMDY]
+    """
 
-    This routine returns two tuples:
-    The first is of (CRVAL1, CRVAL2, PA_y_axis)
-    The second is of (V1ra, V1dec, V3pa)
-    All angles are in decimal degrees.
-    '''
+    # Calculate transforms
+    tforms = calc_transforms(pointing, siaf)
 
-    q1, q2, q3, q4 = q
-    m_eci2j = np.array(
-        [[1. - 2.*q2*q2 - 2.*q3*q3,
-            2.*(q1*q2 + q3*q4),
-            2.*(q3*q1 - q2*q4)],
-         [2.*(q1*q2 - q3*q4),
-            1. - 2.*q3*q3 - 2.*q1*q1,
-            2.*(q2*q3 + q1*q4)],
-         [2.*(q3*q1 + q2*q4),
-            2.*(q2*q3 - q1*q4),
-            1. - 2.*q1*q1 - 2.*q2*q2]])
+    # Calculate the V1 WCS information
+    vinfo = calc_v1_wcs(tforms.m_eci2v)
 
-    mj2fgs1 = np.array(j2fgs_matrix).reshape((3, 3)).transpose()
+    # Calculate the Aperture WCS
+    wcsinfo = calc_aperture_wcs(tforms.m_eci2siaf)
 
-    m_sifov_fsm_delta = np.array(
-        [[1., fsmcorr[0]/22.01, fsmcorr[1]/21.68],
-         [-fsmcorr[0]/22.01, 1., 0.],
-         [-fsmcorr[1]/21.68, 0., 1.]])
+    # That's all folks
+    return (wcsinfo, vinfo)
 
-    mpartial = np.dot(m_sifov_to_v,
-                      np.dot(m_sifov_fsm_delta,
-                             np.dot(m_fgs1_to_sifov, mj2fgs1)))
 
-    m_eci2v = np.dot(mpartial, m_eci2j)
+def calc_transforms(pointing, siaf):
+    """Calculate transforms from pointing to SIAF
 
-    v1pt = m_eci2v[0]
-    xeci_ra, xeci_dec = vector_to_ra_dec(v1pt)
+    Given the spacecraft pointing parameters and the
+    aperture-specific SIAF, calculate all the transforms
+    necessary to produce WCS information.
 
-    # V1 is given by the first row.
-    v1_ra, v1_dec = vector_to_ra_dec(m_eci2v[0])
-    # V3 is given by the third row
-    v3_ra, v3_dec = vector_to_ra_dec(m_eci2v[2])
+    Parameters
+    ----------
+    pointing: Pointing
+        Observatory pointing information
 
-    # The V3PA @ V1 is given by
-    y = cos(v3_dec) * sin(v3_ra-v1_ra)
-    x = sin(v3_dec) * cos(v1_dec) - \
-        cos(v3_dec) * sin(v1_dec) * cos((v3_ra - v1_ra))
-    V3PA = np.arctan2(y, x)
+    siaf: SIAF
+        Aperture information
 
-    m_eci2siaf = np.dot(m_v_to_siaf(v3idlyang * D2R,
-                                    v3ref * A2R,
-                                    v2ref * A2R,
-                                    vidlparity), m_eci2v)
+    Returns
+    -------
+    transforms: Transforms
+        The list of coordinate matrix transformations
+
+    Notes
+    -----
+    The matrix transform pipeline to convert from ECI J2000 observatory
+    qauternion pointing to aperture ra/dec/roll information
+    is given by the following formula. Each term is a 3x3 matrix:
+
+        M_eci_to_siaf =           # The complete transformation
+            M_v1_to_siaf      *   # V1 to SIAF
+            M_sifov_to_v1     *   # Science Instruments Aperture to V1
+            M_sifov_fsm_delta *   # Fine Steering Mirror correction
+            M_fgs1_to_sifov   *   # FGS1 to Science Instruments Aperture
+            M_j_to_fgs1       *   # J-Frame to FGS1
+            M_eci_to_j        *   # ECI to J-Frame
+    """
+
+    tforms = Transforms()
+
+    # Determine the ECI to J-frame matrix
+    tforms.m_eci2j = calc_eci2j_matrix(pointing.q)
+
+    # Calculate the J-frame to FGS! ICS matrix
+    tforms.m_j2fgs1 = calc_j2fgs1_matrix(pointing.j2fgs_matrix)
+
+    # Calculate the FSM corrections to the SI_FOV frame
+    tforms.m_sifov_fsm_delta = calc_sifov_fsm_delta_matrix(pointing.fsmcorr)
+
+    # Calculate the FGS1 ICS to SI-FOV matrix
+    tforms.m_fgs12sifov = calc_fgs1_to_sifov_mastrix()
+
+    # Calculate ECI to SI FOV
+    tforms.m_eci2sifov = np.dot(
+        tforms.m_sifov_fsm_delta,
+        np.dot(
+            tforms.m_fgs12sifov,
+            np.dot(
+                tforms.m_j2fgs1,
+                tforms.m_eci2j
+            )
+        )
+    )
+
+    # Calculate SI FOV to V1 matrix
+    tforms.m_sifov2v = calc_sifov2v_matrix()
+
+    # Calculate the complete transform to the V1 reference
+    tforms.m_eci2v = np.dot(
+        tforms.m_sifov2v,
+        tforms.m_eci2sifov
+    )
+
+    # Calculate the SIAF transform matrix
+    tforms.m_v2siaf = calc_v2siaf_matrix(siaf)
+
+    # Calculate the full ECI to SIAF transform matrix
+    tforms.m_eci2siaf = np.dot(
+        tforms.m_v2siaf,
+        tforms.m_eci2v
+    )
+
+    return tforms
+
+
+def calc_v1_wcs(m_eci2v):
+    """Calculate the V1 wcs information
+
+    Parameters
+    ----------
+    m_eci2v: np.array((3, 3))
+        The ECI to V transformation matrix
+
+    Returns
+    -------
+    vinfo: WCSRef
+        The V1 wcs pointing
+    """
+    vinfo = WCSRef()
+
+    # V1 RA/Dec is the first row of the transform
+    vinfo.ra, vinfo.dec = vector_to_ra_dec(m_eci2v[0])
+
+    # V3 is the third row of the transformation
+    v3info = WCSRef()
+    v3info.ra, v3info.dec = vector_to_ra_dec(m_eci2v[2])
+
+    # Calculate the V3 position angle
+    vinfo.pa = calc_position_angle(vinfo, v3info)
+
+    # Convert to degrees
+    vinfo = WCSRef(
+        ra=vinfo.ra * R2D,
+        dec=vinfo.dec * R2D,
+        pa=vinfo.pa * R2D
+    )
+
+    return vinfo
+
+
+def calc_aperture_wcs(m_eci2siaf):
+    """Calculate the aperture WCS
+
+    Parameters
+    ----------
+    m_eci2siaf: np.array((3, 3))
+        The ECI to SIAF transformation matrix
+
+    Returns
+    -------
+    wcsinfo: WCSRef
+        The aperturn wcs information
+    """
+
+    # Calculate point on sky.
+    # Note, the SIAF referenct point is hardcoded to
+    # (0, 0). The calculation is left here in case
+    # this is not desired.
+    wcsinfo = WCSRef()
     siaf_x = 0. * A2R
     siaf_y = 0. * A2R
     refpos = np.array(
-               [siaf_x,
-                siaf_y,
-                np.sqrt(1.-siaf_x * siaf_x - siaf_y * siaf_y)])
+        [siaf_x,
+         siaf_y,
+         np.sqrt(1.-siaf_x * siaf_x - siaf_y * siaf_y)]
+    )
     msky = np.dot(m_eci2siaf.transpose(), refpos)
-    vaper_ra, vaper_dec = vector_to_ra_dec(msky)
+    wcsinfo.ra, wcsinfo.dec = vector_to_ra_dec(msky)
 
+    # Calculate the position angle
     vysiaf = np.array([0., 1., 0.])
-
     myeci = np.dot(m_eci2siaf.transpose(), vysiaf)
+
     # The Y axis of the aperture is given by
     vy_ra, vy_dec = vector_to_ra_dec(myeci)
+
     # The VyPA @ xref,yref is given by
-    y = cos(vy_dec) * sin(vy_ra-vaper_ra)
-    x = sin(vy_dec) * cos(vaper_dec) - \
-        cos(vy_dec) * sin(vaper_dec) * cos((vy_ra - vaper_ra))
-    vypa = np.arctan2(y, x)
-    wcsinfo = (vaper_ra*R2D, vaper_dec*R2D, vypa*R2D)
-    vinfo = (v1_ra*R2D, v1_dec*R2D, V3PA*R2D)
-    return wcsinfo, vinfo
+    y = cos(vy_dec) * sin(vy_ra-wcsinfo.ra)
+    x = sin(vy_dec) * cos(wcsinfo.dec) - \
+        cos(vy_dec) * sin(wcsinfo.dec) * cos((vy_ra - wcsinfo.ra))
+    wcsinfo.pa = np.arctan2(y, x)
+
+    # Convert all WCS to degrees
+    wcsinfo = WCSRef(
+        ra=wcsinfo.ra * R2D,
+        dec=wcsinfo.dec * R2D,
+        pa=wcsinfo.pa * R2D
+    )
+    
+    return wcsinfo
 
 
-def get_pointing(obstart, obsend, result_type='first'):
+def calc_eci2j_matrix(q):
+    """Calculate ECI to J-frame matrix from quaternions
+
+    Parameters
+    ----------
+    q: np.array(q1, q2, q3, q4)
+        Array of quaternions from the engineering database
+
+    Returns
+    -------
+    transform: np.array((3, 3))
+        The transform matrix representing the transformation
+        from observatory orientation to J-Frame
+    """
+    q1, q2, q3, q4 = q
+    transform = np.array(
+        [[1. - 2.*q2*q2 - 2.*q3*q3,
+          2.*(q1*q2 + q3*q4),
+          2.*(q3*q1 - q2*q4)],
+         [2.*(q1*q2 - q3*q4),
+          1. - 2.*q3*q3 - 2.*q1*q1,
+          2.*(q2*q3 + q1*q4)],
+         [2.*(q3*q1 + q2*q4),
+          2.*(q2*q3 - q1*q4),
+          1. - 2.*q1*q1 - 2.*q2*q2]]
+    )
+
+    return transform
+
+
+def calc_j2fgs1_matrix(j2fgs_matrix):
+    """Calculate the J-frame to FGS1 transformation
+
+    Parameters
+    ----------
+    j2fgs_matrix: n.array((9,))
+        Matrix parameters from the engineering database.
+        If all zeros, a predefined matrix is used.
+
+    Returns
+    -------
+    transform: np.array((3, 3))
+        The transformation matrix
+    """
+    if np.isclose(j2fgs_matrix, 0.).all():
+        logger.warning(
+            'J-Frame to FGS1 engineering parameters are all zero.'
+            '\nUsing default matrix'
+        )
+        m_partial = np.asarray(
+            [
+                [0., 1., 0.],
+                [0., 0., 1.],
+                [1., 0., 0.]
+            ]
+        )
+        transform = np.dot(
+            m_partial,
+            J2FGS_MATRIX_DEFAULT
+        )
+
+    else:
+        logger.info(
+            'Using J-Frame to FGS1 engineering parameters'
+            'for the J-Frame to FGS1 transformation.'
+        )
+        transform = np.array(j2fgs_matrix).reshape((3, 3)).transpose()
+
+    return transform
+
+
+def calc_sifov_fsm_delta_matrix(fsmcorr):
+    """Calculate Fine Steering Mirror correction matrix
+
+    Parameters
+    ----------
+    fsmcorr: np.array((2,))
+        The FSM correction parameters
+
+    Returns
+    -------
+    transform: np.array((3, 3))
+        The transformation matrix
+    """
+    transform = np.array(
+        [
+            [1.,                fsmcorr[0]/22.01, fsmcorr[1]/21.68],
+            [-fsmcorr[0]/22.01,               1.,               0.],
+            [-fsmcorr[1]/21.68,               0.,               1.]
+        ]
+    )
+
+    return transform
+
+
+def calc_fgs1_to_sifov_mastrix():
+    """
+    Calculate the FGS! to SI-FOV matrix
+
+    Currently, this is a defined matrix
+    """
+    m_partial = np.array(
+        [[0, 0, 1],
+         [1, 0, 0],
+         [0, 1, 0]]
+    )
+
+    transform = np.dot(m_partial, FGS12SIFOV_DEFAULT)
+    return transform
+
+
+def calc_sifov2v_matrix():
+    """Calculate the SI-FOV to V-Frame matrix
+
+    This is currently defined as the inverse Euler rotation
+    about an angle of 7.8 arcmin. Here returns the pre-calculate
+    matrix.
+    """
+    return SIFOV2V_DEFAULT
+
+
+def calc_v2siaf_matrix(siaf):
+    """Calculate the SIAF transformation matrix
+
+    Parameters
+    ----------
+    siaf: SIAF
+        The SIAF parameters
+
+    Returns
+    -------
+    transform: np.array((3, 3))
+        The V1 to SIAF transformation matrix
+    """
+    v2, v3, v3idlyang, vparity = siaf
+    v2 *= A2R
+    v3 *= A2R
+    v3idlyang *= D2R
+
+    mat = np.array(
+        [[cos(v3)*cos(v2),
+          cos(v3)*sin(v2),
+          sin(v3)],
+         [-cos(v3idlyang)*sin(v2)+sin(v3idlyang)*sin(v3)*cos(v2),
+          cos(v3idlyang)*cos(v2)+sin(v3idlyang)*sin(v3)*sin(v2),
+          -sin(v3idlyang)*cos(v3)],
+         [-sin(v3idlyang)*sin(v2)-cos(v3idlyang)*sin(v3)*cos(v2),
+          sin(v3idlyang)*cos(v2)-cos(v3idlyang)*sin(v3)*sin(v2),
+          cos(v3idlyang)*cos(v3)]])
+    pmat = np.array([[0., vparity, 0.],
+                     [0., 0., 1.],
+                     [1., 0., 0.]])
+
+    transform = np.dot(pmat, mat)
+    return transform
+
+
+def calc_position_angle(v1, v3):
+    """Calculate V3 position angle @V1
+
+    Parameters
+    ----------
+    v1: WCSRef
+        The V1 wcs parameters
+
+    v3: WCSRef
+        The V3 wcs parameters
+
+    Returns
+    -------
+    v3_pa: float
+      The V3 position angle, in radians
+    """
+    y = cos(v3.dec) * sin(v3.ra-v1.ra)
+    x = sin(v3.dec) * cos(v1.dec) - \
+        cos(v3.dec) * sin(v1.dec) * cos((v3.ra - v1.ra))
+    v3_pa = np.arctan2(y, x)
+
+    return v3_pa
+
+
+def get_pointing(obsstart, obsend, result_type='first'):
     """
     Get telescope pointing engineering data.
 
     Parameters
     ----------
-    obstart, obsend: float
+    obsstart, obsend: float
         MJD observation start/end times
 
     result_type: str
@@ -327,10 +609,10 @@ def get_pointing(obstart, obsend, result_type='first'):
 
     Returns
     -------
-    q, j2fgs_matrix, fsmcorr, obstime
+    pointing: Pointing or [Pointing(, ...)]
         The engineering pointing parameters.
-        If the `result_type` returns multiple values, what is returned
-        will be a list of 4-tuples.
+        If the `result_type` is `all`, a list
+        of pointings will be returned
 
     Raises
     ------
@@ -345,8 +627,8 @@ def get_pointing(obstart, obsend, result_type='first'):
     """
     logger.info(
         'Determining pointing between observations times (mjd):'
-        '\n\tobstart = {}'
-        '\n\tobsend = {}'.format(obstart, obsend)
+        '\n\tobsstart = {}'
+        '\n\tobsend = {}'.format(obsstart, obsend)
     )
     logger.info(
         'Querying engineering DB: {}'.format(ENGDB_BASE_URL)
@@ -380,7 +662,7 @@ def get_pointing(obstart, obsend, result_type='first'):
     for param in params:
         try:
             params[param] = engdb.get_values(
-                param, obstart, obsend, time_format='mjd', include_obstime=True
+                param, obsstart, obsend, time_format='mjd', include_obstime=True
             )
         except Exception as exception:
             raise ValueError(
@@ -399,19 +681,20 @@ def get_pointing(obstart, obsend, result_type='first'):
             for param in params
         ]
         if any(values):
+            pointing = Pointing()
 
             # The tagged obstime will come from the SA_ZATTEST1 mneunonic
-            obstime = params['SA_ZATTEST1'][idx].obstime
+            pointing.obstime = params['SA_ZATTEST1'][idx].obstime
 
             # Fill out the matricies
-            q = np.array([
+            pointing.q = np.array([
                 params['SA_ZATTEST1'][idx].value,
                 params['SA_ZATTEST2'][idx].value,
                 params['SA_ZATTEST3'][idx].value,
                 params['SA_ZATTEST4'][idx].value,
             ])
 
-            j2fgs_matrix = np.array([
+            pointing.j2fgs_matrix = np.array([
                 params['SA_ZRFGS2J11'][idx].value,
                 params['SA_ZRFGS2J21'][idx].value,
                 params['SA_ZRFGS2J31'][idx].value,
@@ -423,18 +706,13 @@ def get_pointing(obstart, obsend, result_type='first'):
                 params['SA_ZRFGS2J33'][idx].value,
             ])
 
-            fsmcorr = np.array([
+            pointing.fsmcorr = np.array([
                 params['SA_ZADUCMDX'][idx].value,
                 params['SA_ZADUCMDY'][idx].value,
 
             ])
 
-            results.append(Pointing_Quaternions(
-                q=q,
-                j2fgs_matrix=j2fgs_matrix,
-                fsmcorr=fsmcorr,
-                obstime=obstime
-            ))
+            results.append(pointing)
 
             # Short circuit if all we're looking for is the first.
             if result_type == 'first':
@@ -443,7 +721,7 @@ def get_pointing(obstart, obsend, result_type='first'):
     if not len(results):
         raise ValueError(
                 'No non-zero quanternion found '
-                'in the DB between MJD {} and {}'.format(obstart, obsend)
+                'in the DB between MJD {} and {}'.format(obsstart, obsend)
             )
 
     if result_type == 'first':
@@ -452,33 +730,29 @@ def get_pointing(obstart, obsend, result_type='first'):
         return results
 
 
-def get_pointing_stub(obstart, obsend):
-    '''
-    For the time being this simply returns the same damn values regardless of
-    the input time (awaiting the time that these parameters are actually
-    in the engineering database)
-    '''
-    # The following values of q correspond to the engineering keyword values:
-    # SA_ZATEST1, SA_ZATEST2, SA_ZATEST3, SA_ZATEST4
-    q = np.array([-0.36915286,  0.33763282,  0.05758533,  0.86395264])
-    # The following values of j2fgs_matrix correspond to the engineering
-    #  keyword values of:
-    # SA_ZRFGS2K11 SA_ZRFGS2K21 SA_ZRFGS2K31
-    # SA_ZRFGS2K21 SA_ZRFGS2K22 SA_ZRFGS2K32
-    # SA_ZRFGS2K31 SA_ZRFGS2K32 SA_ZRFGS2K33
-    j2fgs_matrix = np.array(
-        [-1.00444000e-03,  3.38145836e-03, 9.99993778e-01,
-         9.99999496e-01, -3.90000000e-14, 1.00444575e-03,
-         3.39649146e-06,  9.99994283e-01, -3.38145665e-03])
-    # The following values of fsmcorr correspond to the engineering keywords:
-    # SA_ZADUCMDX, SA_ZADUCMDY
-    fsmcorr = np.array([0., 0.])
-    return q, j2fgs_matrix, fsmcorr
+def vector_to_ra_dec(v):
+    """Returns tuple of spherical angles from unit direction Vector
+
+    Parameters
+    ----------
+    v: [v0, v1, v2]
+
+    Returns
+    -------
+    ra, dec: float, float
+        The spherical angles, in radians
+    """
+    ra = np.arctan2(v[1], v[0])
+    dec = np.arcsin(v[2])
+    if ra < 0.:
+        ra += 2. * np.pi
+    return(ra, dec)
 
 
 def compute_local_roll(pa_v3, ra_ref, dec_ref, v2_ref, v3_ref):
     """
-    Computes the position angle of V3 (measured N to E) at the center af an aperture.
+    Computes the position angle of V3 (measured N to E)
+    at the center af an aperture.
 
     Parameters
     ----------
@@ -501,16 +775,17 @@ def compute_local_roll(pa_v3, ra_ref, dec_ref, v2_ref, v3_ref):
     dec_ref = np.deg2rad(dec_ref)
     pa_v3 = np.deg2rad(pa_v3)
 
-    M = np.array([[cos(ra_ref) * cos(dec_ref),
-                   -sin(ra_ref) * cos(pa_v3) + cos(ra_ref) * sin(dec_ref) * sin(pa_v3),
-                   -sin(ra_ref) * sin(pa_v3) - cos(ra_ref) * sin(dec_ref) * cos(pa_v3)],
-                  [sin(ra_ref) * cos(dec_ref),
-                   cos(ra_ref) * cos(pa_v3) + sin(ra_ref) * sin(dec_ref) * sin(pa_v3),
-                   cos(ra_ref) * sin(pa_v3) - sin(ra_ref) * sin(dec_ref) * cos(pa_v3)],
-                   [sin(dec_ref),
-                    -cos(dec_ref) * sin(pa_v3),
-                    cos(dec_ref) * cos(pa_v3)]
-                  ])
+    M = np.array(
+        [[cos(ra_ref) * cos(dec_ref),
+          -sin(ra_ref) * cos(pa_v3) + cos(ra_ref) * sin(dec_ref) * sin(pa_v3),
+          -sin(ra_ref) * sin(pa_v3) - cos(ra_ref) * sin(dec_ref) * cos(pa_v3)],
+         [sin(ra_ref) * cos(dec_ref),
+          cos(ra_ref) * cos(pa_v3) + sin(ra_ref) * sin(dec_ref) * sin(pa_v3),
+          cos(ra_ref) * sin(pa_v3) - sin(ra_ref) * sin(dec_ref) * cos(pa_v3)],
+         [sin(dec_ref),
+          -cos(dec_ref) * sin(pa_v3),
+          cos(dec_ref) * cos(pa_v3)]]
+    )
 
     return _roll_angle_from_matrix(M, v2, v3)
